# openapi.yaml
openapi: "3.0.3"
info:
  title: Shepherd API
  version: "0.1.0"

paths:
  /api/v1/health:
    get:
      operationId: getHealth
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                required: [status]

  /api/v1/ws:
    get:
      operationId: listWorkspaces
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  workspaces:
                    type: array
                    items:
                      $ref: "#/components/schemas/Workspace"
                required: [workspaces]

  /api/v1/ws/{workspaceId}:
    get:
      operationId: getWorkspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "404":
          description: Not Found

  /api/v1/ws/{workspaceId}/config:
    get:
      operationId: getWorkspaceConfig
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceConfig"
        "404":
          description: Not Found

  /api/v1/ws/{workspaceId}/tickets:
    get:
      operationId: listTickets
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: statusId
          in: query
          schema:
            type: string
        - name: isClosed
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  tickets:
                    type: array
                    items:
                      $ref: "#/components/schemas/Ticket"
                required: [tickets]
        "404":
          description: Not Found
    post:
      operationId: createTicket
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "404":
          description: Not Found

  /api/v1/ws/{workspaceId}/tickets/{ticketId}:
    get:
      operationId: getTicket
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/TicketId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "404":
          description: Not Found
    patch:
      operationId: updateTicket
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/TicketId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTicketRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "404":
          description: Not Found
    delete:
      operationId: deleteTicket
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/TicketId"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not Found

  /api/v1/ws/{workspaceId}/slack/users:
    get:
      operationId: listSlackUsers
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/SlackUserInfo"
                required: [users]
        "404":
          description: Not Found

  /api/v1/ws/{workspaceId}/slack/users/{userId}:
    get:
      operationId: getSlackUserInfo
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlackUserInfo"
        "404":
          description: Not Found

  /api/v1/ws/{workspaceId}/sources:
    get:
      operationId: listSources
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  sources:
                    type: array
                    items:
                      $ref: "#/components/schemas/Source"
                required: [sources]
    post:
      operationId: createSource
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSourceRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"
        "400":
          description: Bad Request
        "409":
          description: Conflict

  /api/v1/ws/{workspaceId}/sources/{sourceId}:
    delete:
      operationId: deleteSource
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: sourceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found

  /api/v1/ws/{workspaceId}/tools:
    get:
      operationId: listToolSettings
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  tools:
                    type: array
                    items:
                      $ref: "#/components/schemas/ToolState"
                required: [tools]

  /api/v1/ws/{workspaceId}/tools/{providerId}:
    put:
      operationId: setToolEnabled
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: providerId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                enabled:
                  type: boolean
              required: [enabled]
      responses:
        "204":
          description: No Content

  /api/v1/ws/{workspaceId}/tickets/{ticketId}/comments:
    get:
      operationId: listComments
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/TicketId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  comments:
                    type: array
                    items:
                      $ref: "#/components/schemas/Comment"
                required: [comments]
        "404":
          description: Not Found

components:
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
    TicketId:
      name: ticketId
      in: path
      required: true
      schema:
        type: string

  schemas:
    Workspace:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      required: [id, name]

    StatusDef:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
        order:
          type: integer
        isClosed:
          type: boolean
      required: [id, name, color, order, isClosed]

    TicketConfig:
      type: object
      properties:
        defaultStatusId:
          type: string
        closedStatusIds:
          type: array
          items:
            type: string
      required: [defaultStatusId, closedStatusIds]

    FieldType:
      type: string
      enum: [text, number, select, multi-select, user, multi-user, date, url]

    FieldOption:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
        metadata:
          type: object
          additionalProperties: true
      required: [id, name]

    FieldDefinition:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          $ref: "#/components/schemas/FieldType"
        required:
          type: boolean
        description:
          type: string
        options:
          type: array
          items:
            $ref: "#/components/schemas/FieldOption"
      required: [id, name, type, required]

    EntityLabels:
      type: object
      properties:
        ticket:
          type: string
        title:
          type: string
        description:
          type: string
      required: [ticket, title, description]

    WorkspaceConfig:
      type: object
      properties:
        statuses:
          type: array
          items:
            $ref: "#/components/schemas/StatusDef"
        ticketConfig:
          $ref: "#/components/schemas/TicketConfig"
        fields:
          type: array
          items:
            $ref: "#/components/schemas/FieldDefinition"
        labels:
          $ref: "#/components/schemas/EntityLabels"
      required: [statuses, ticketConfig, fields, labels]

    FieldValue:
      type: object
      properties:
        fieldId:
          type: string
          x-go-type: types.FieldID
          x-go-type-import:
            path: github.com/m-mizutani/shepherd/pkg/domain/types
        value: {}
      required: [fieldId, value]

    Ticket:
      type: object
      properties:
        id:
          type: string
        seqNum:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        statusId:
          type: string
        assigneeId:
          type: string
        reporterSlackUserId:
          type: string
        slackChannelId:
          type: string
        slackThreadTs:
          type: string
        fields:
          type: array
          items:
            $ref: "#/components/schemas/FieldValue"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, seqNum, title, statusId, fields, createdAt, updatedAt]

    CreateTicketRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        statusId:
          type: string
        assigneeId:
          type: string
        fields:
          type: array
          items:
            $ref: "#/components/schemas/FieldValue"
      required: [title]

    UpdateTicketRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        statusId:
          type: string
        assigneeId:
          type: string
        fields:
          type: array
          items:
            $ref: "#/components/schemas/FieldValue"

    SlackUserInfo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        imageUrl:
          type: string
      required: [id, name]

    Comment:
      type: object
      properties:
        id:
          type: string
        slackUserId:
          type: string
        body:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, slackUserId, body, createdAt]

    NotionSource:
      type: object
      properties:
        objectType:
          type: string
          enum: [page, database]
        objectId:
          type: string
        url:
          type: string
        title:
          type: string
      required: [objectType, objectId, url]

    Source:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        provider:
          type: string
          enum: [notion]
        notion:
          $ref: "#/components/schemas/NotionSource"
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
      required: [id, workspaceId, provider, createdAt]

    CreateSourceRequest:
      type: object
      properties:
        provider:
          type: string
          enum: [notion]
        url:
          type: string
      required: [provider, url]

    ToolState:
      type: object
      properties:
        providerId:
          type: string
        available:
          type: boolean
        defaultEnabled:
          type: boolean
        enabled:
          type: boolean
        reason:
          type: string
          enum: [provider_unavailable, workspace_disabled, gate_blocked]
      required: [providerId, available, defaultEnabled, enabled]